nova/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.1.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.2
platformdirs==4.2.1
psycopg2-binary==2.9.10
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
